+++ src/classes/StoredItem.cs
@@ -106,6 +106,7 @@
         Sacrificer,
         SurfacedDieItem,
         Rusty,
+        BabyKiwiEgg,
     }